Init with Spring Boot Web, Spring Boot JDBC, and MySQL JDBC Connector
We are going to start with a simple Spring Boot application generated via Spring Initializr.
This is a demonstration application which does two things: fetches some data from the relational
database and exposes this information via an HTTP endpoint. A pretty typical web application setup.

In real life, there could be more moving pieces and dependencies like message brokers and ORMs but
this minimal setup should be sufficient to showcase the core concepts and techniques that will allow
to fully modularize a typical Java application.