Add a demonstration controller
This is our very simple business logic for today. A @RestController + plain JDBC. The result is a record which is serialized by Jackson.

For convenience, this commit adds the `docker-compose.yml` with the MySQL Server configuration, the two IJ IDEA run configurations (one for starting the application, the other is for introspecting the runtime classpath), and the test script for playing with our HTTP API.

Let's run our app and inspect it with VisualVM.