Avoid usage of the classpath
Even though the module system can work with JARs that carry no module descriptor (no Automatic-Module-Name in Manifest and no `module-info.class`), Gradle does not do this and always places such JARs on the classpath.

https://github.com/gradle/gradle/issues/890#issuecomment-603509389

Let's now unleash the power of Gradle and try to overcome this problem by using Artifact Transform.

https://docs.gradle.org/current/userguide/artifact_transforms.html
https://github.com/gradlex-org/extra-java-module-info

The benefit of using this abstraction is that it is completely transparent to the tooling (IJ, application plugin, etc).