Define module-info.class for the automatic module dependencies
This metadata was generated by the moduleDescriptorRecommendations task ( extra-java-module-info plugin).

This is a suboptimal yet almost working configuration. Ideally, the module definitions should be refined, e.g. one would need to convert some of the "requires"/"requires transitive" to "requires static". The majority of the "org.springframework.boot:spring-boot-autoconfigure" module "requires" should become "requires static" while "org.springframework.boot:spring-boot-starter-*" should be amended with its direct dependencies.

The downside of this approach is its maintainability. It works very well for the old libraries that no longer receive updates, but it is very fragile when it comes to patching libraries that are being actively maintained, i.e. Spring Boot. You, effectively, become a maintainer of the library who is now responsible for updating this module metadata.

The good news is that with every new release of Spring Boot, more and more libraries become proper modules.